+++ packages/react-router/modules/withRouter.js
@@ -12,6 +12,7 @@ const withRouter = (Component) => {
   )
 
   C.displayName = `withRouter(${Component.displayName || Component.name})`
+  C.WrappedComponent = Component
 
   return C
 }